Python_HID_Bridge/requirements.txt
pyserial>=3.5
pynput>=1.7.6 